+++ covalent_dispatcher_legacy/tests/cli/service_test.py
@@ -47,7 +47,7 @@ from covalent_dispatcher_legacy._cli.service import (
 )
 
 STOPPED_SERVER_STATUS_ECHO = "Covalent server is stopped.\n"
-RUNNING_SERVER_STATUS_ECHO = "Covalent server is running at http://0.0.0.0:42.\n"
+RUNNING_SERVER_STATUS_ECHO = "Covalent server is running at http://localhost:42.\n"
 
 STARTED_SUPERVISORD_ECHO = "Started Supervisord process {}."
 NO_SERVICE_PROVIDED_TO_LOGS_ECHO = (