+++ CHANGELOG.md
@@ -634,7 +634,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Update `covalent-ui` service in `docker-compose.yaml` to ensure that the uvicorn server listens on `0.0.0.0` for all incoming requests
+- Update `covalent-ui` service in `docker-compose.yaml` to ensure that the uvicorn server listens on `localhost` for all incoming requests
 - Using `ENTRYPOINT` in dockerfiles instead of `CMD`
 - Remove `command` option from all services in `docker-compose.yml`